feat(master): add MITM proxy to modify HTTPS requests
