+++ .gitignore
@@ -3,3 +3,4 @@ node_modules/
 .DS_Store
 server/my-config.json
 server/conf.json
+.http-mitm-proxy/
